Revert "client util split"
This reverts commit dc43366635b90726b116612a8864970b74a3e04c.
